proteomDjango231231/celery_Proteomics/requirements/condaRequirements.txt
amqp==5.1.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.11.2
billiard==3.6.4.0
biopython==1.79
bleach==6.0.0
Bottleneck==1.3.5
celery==5.2.7
celery-progress==0.2
certifi==2022.12.7
cffi==1.15.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
comm==0.1.2
contourpy==1.0.5
cryptography==3.4.8
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Django==3.2
django-admin-import==0.3.0
django-celery-beat==2.2.1
django-celery-results==2.3.1
django-cors-headers==3.14.0
django-environ==0.4.5
django-extensions==3.2.1
django-filter==22.1
django-markdownx==3.0.1
django-schema-graph==2.2.1
django-spaghetti-and-meatballs==0.4.2
django-timezone-field==4.1.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
docopt==0.6.2
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.3
flit_core==3.8.0
fonttools==4.25.0
goatools==1.3.1
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel==6.21.3
ipython==8.11.0
ipython-genutils==0.2.0
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter-client==7.3.4
jupyter_core==5.2.0
jupyter-server==1.23.6
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
kombu==5.2.4
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
mistune==2.0.5
munkres==1.1.4
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==2.8.4
notebook==6.5.3
notebook_shim==0.2.2
numba==0.56.4
numexpr==2.8.4
numpy==1.23.5
openpyxl==3.1.1
packaging==22.0
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.1.0
ply==3.11
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.4
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydot==1.4.2
Pygments==2.14.0
PyJWT==1.7.1
pyparsing==3.0.9
PyQt5-sip==12.11.0
pyrsistent==0.19.3
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
pyzmq==25.0.0
qnorm==0.8.1
requests==2.28.2
rpy2==3.5.1
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.0
setuptools==65.6.3
simplegeneric==0.8.1
sip==6.6.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.2
statsmodels==0.13.5
terminado==0.17.1
tinycss2==1.2.1
toml==0.10.2
tornado==6.1
traitlets==5.9.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.1
wheel==0.38.4
XlsxWriter==3.0.8
zipp==3.15.0
